+++ src/pages/UniversalSwap/Modals/SelectTokenModal.module.scss
