+++ src/pages/UniversalSwap/Modals/SlippageModal.module.scss
